refactor(db): Separacion del archivo seed
